requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.7
Django==4.2.4
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.2
djangorestframework==3.14.0
Faker==20.1.0
fcm-django==2.0.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.24.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.4
msgpack==1.0.7
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
